// src/packages/h5p-server/types.ts
export type ContentId = string;
// ...
export interface ILibraryName {
    /**
     * The name used to identify the library (e.g. H5P.Example)
     */
    machineName: string;
    /**
     * The major version of the library (e.g. 1)
     */
    majorVersion: number;
    /**
     * The minor version of the library (e.g. 0)
     */
    minorVersion: number;
}

export interface IFullLibraryName extends ILibraryName {
    patchVersion: number;
}

export interface ILibraryInstallResult {
    newVersion?: IFullLibraryName;
    oldVersion?: IFullLibraryName;
    type: 'new' | 'patch' | 'none';
}

/**
 * This is an author inside content metadata.
 */
export interface IContentAuthor {
    name?: string;
    role?: string;
}

/**
 * This is a change inside content metadata.
 */
export interface IContentChange {
    author?: string;
    date?: string;
    log?: string;
}

/**
 * This is the structure of the object received by the editor client when saving content.
 * It is also used when creating h5p.json files for .h5p packages.
 */
export interface IContentMetadata {
    a11yTitle?: string;
    author?: string;
    authorComments?: string;
    authors?: IContentAuthor[];
    changes?: IContentChange[];
    contentType?: string;
    dynamicDependencies?: ILibraryName[];
    editorDependencies?: ILibraryName[];
    embedTypes: ('iframe' | 'div')[];
    h?: string;
    language: string;
    license?: string;
    licenseExtras?: string;
    licenseVersion?: string;
    mainLibrary: string;
    metaDescription?: string;
    metaKeywords?: string;
    /**
     * A flat list of all dependencies required for the content to be displayed.
     * Note that this list is a flat representation of the full dependency tree
     * that is created by looking up all dependencies of the main library.
     */
    preloadedDependencies: ILibraryName[];
    source?: string;
    title: string;
    w?: string;
    yearsFrom?: string;
    yearsTo?: string;
}
// ...
export interface IUser {
    /**
     * If true, the user can create content of content types that are set to "restricted".
     */
    canCreateRestricted: boolean;
    /**
     * If true, the user can install content types from the hub that are set the "recommended"
     * by the Hub.
     */
    canInstallRecommended: boolean;
    /**
     * If true, the user can generally install and update libraries. This includes Hub
     * content types that aren't set to "recommended" or uploading custom packages.
     */
    canUpdateAndInstallLibraries: boolean;
    /**
     * E-Mail address.
     */
    email: string;
    /**
     * An internal id used to check if user objects are identical.
     */
    id: string;
    /**
     * The full name of the user.
     */
    name: string;
    /**
     * Specifies type of user. Possible values other than 'local' are unknown as of this time.
     */
    type: 'local' | string;
}
// ...
export interface IEditorModel {
    integration: IIntegration;
    scripts: string[];
    styles: string[];
    urlGenerator: IUrlGenerator;
}

export interface IUrlGenerator {
    ajaxEndpoint(user: IUser): string;
    baseUrl(): string;
    /**
     * Generates a URL at which the resources of the content can be loaded.
     * Should be undefined if the default content serving mechanism is used. You
     * only have to return a URL here if you want to change the hostname of the
     * route, e.g. if the content files can be requested directly from S3 or a
     * CDN.
     * @param contentId
     * @returns the URL or undefined (Example:
     * http://127.0.0.1:9000/s3bucket/123`)
     */
    contentFilesUrl(contentId: ContentId): string | undefined;
    contentUserData(user: IUser): string;
    coreFile(file: string): string;
    coreFiles(): string;
    downloadPackage(contentId: ContentId): string;
    editorLibraryFile(file: string): string;
    editorLibraryFiles(): string;
    libraryFile(library: IFullLibraryName, file: string): string;
    parameters(): string;
    play(): string;
    setFinished(user: IUser): string;
    temporaryFiles(): string;
    /**
     * Generates a URL by which the content can be globally identified (or
     * accessed). The URL is used when generating ids for object in xAPI
     * statements, so it could also be a fake URL that is only used to uniquely
     * identify the content. It could also simply be the contentId. (default
     * behavior) If you store the generated xAPI statement in a neutral LRS and
     * users will see or be able to click the URL, you should customize the URL
     * by overriding this method.
     * @param contentId
     */
    uniqueContentUrl(contentId: ContentId): string;
}